upgrade and optimise homework15

+++ src/homework/homework15/medicalCenter/Demo.java
@@ -11,10 +11,13 @@ import java.util.Scanner;
 
 public class Demo implements Commands {
 
-    static Scanner sc = new Scanner(System.in);
-    static Storage storage = new Storage();
-    static DateUtil du = new DateUtil();
-    static Profession[] values = Profession.values();
+    static private final Scanner sc = new Scanner(System.in);
+    static private final Storage storage = new Storage();
+    static private final Profession[] values = Profession.values();
+
+    private static boolean isValidEmail(String email) {
+        return email.contains("@") && (email.endsWith(".ru") || email.endsWith(".com"));
+    }
 
     public static void main(String[] args) throws ParseException {
 
@@ -30,23 +33,39 @@ public class Demo implements Commands {
                     addDoctor();
                     break;
                 case SEARCH_DOCTOR_BY_PROFESSION:
-                    searchDoctorByProfession();
+                    if(storage.getSize() == 0) {
+                        System.out.println("There is no doctor!");
+                    }else {
+                        searchDoctorByProfession();
+                    }
                     break;
                 case DELETE_DOCTOR_BY_ID:
-                    storage.printDoctors();
-                    System.out.println("Choose doctor!");
-                    System.out.print("input id: ");
-                    String delId = sc.nextLine();
-                    storage.deleteDoctorById(delId);
+                    if(storage.getSize() == 0) {
+                        System.out.println("There is no doctor!");
+                    }else {
+                        storage.printDoctors();
+                        System.out.println("Choose doctor!");
+                        System.out.print("input id: ");
+                        String delId = sc.nextLine();
+                        storage.deleteDoctorById(delId);
+                    }
                     break;
                 case CHANGE_DOCTOR_DATA_BY_ID:
-                    changeDoctorDataById();
+                    if(storage.getSize() == 0) {
+                        System.out.println("There is no doctor!");
+                    }else{
+                        changeDoctorDataById();
+                    }
                     break;
                 case ADD_PATIENT:
                     addPatient();
                     break;
                 case PRINT_ALL_PATIENTS_BY_DOCTOR:
-                    printAllPatientsByDoctor();
+                    if(storage.getSize() == 0) {
+                        System.out.println("There is no doctor!");
+                    }else {
+                        printAllPatientsByDoctor();
+                    }
                     break;
                 case PRINT_TODAYS_PATIENTS:
                     storage.printTodaysPatients();
@@ -63,25 +82,29 @@ public class Demo implements Commands {
         System.out.println("Choose Doctor!");
         System.out.print("input doctor's id: ");
         String doctorID = sc.nextLine();
-        Doctor doctor = storage.getDoctorById(doctorID);
-        if(doctor == null){
-            System.out.println("Doctor with id " + doctorID + " not found");
-        }else{
-            storage.printAllPatientsByDoctor(doctor);
+        Doctor doctor;
+
+        try {
+            doctor = storage.getDoctorById(doctorID);
+        } catch (NullPointerException e) {
+            System.out.println("Doctor with id " + doctorID + " not found!");
+            return;
         }
+
+        storage.printAllPatientsByDoctor(doctor);
     }
 
-    private static void addPatient(){
+    private static void addPatient() {
 
         if (storage.getSize() == 0) {
-            System.out.println("There is no doctor");
+            System.out.println("There is no doctor!");
         } else {
-
             storage.printDoctors();
             System.out.println("Choose doctor!");
             System.out.print("input doctor's id: ");
             String patientsDoctorId = sc.nextLine();
             Doctor patientsDoctor = storage.getDoctorById(patientsDoctorId);
+
             if (patientsDoctor == null) {
                 System.out.println("Doctor with id " + patientsDoctorId + " not found");
             } else {
@@ -94,34 +117,36 @@ public class Demo implements Commands {
                 System.out.print("input email: ");
                 String patientEmail = sc.nextLine();
 
-                boolean ec = emailChecker(patientEmail);
+                boolean ec = isValidEmail(patientEmail);
                 while (!ec) {
                     System.out.println("wrong email format, try again! (ex@mail.ru, ex@gmail.com");
                     patientEmail = sc.nextLine();
-                    ec = emailChecker(patientEmail);
+                    ec = isValidEmail(patientEmail);
                 }
 
                 System.out.print("input phone number: ");
                 String patientPhoneNumber = sc.nextLine();
-                
+
                 Date date;
                 System.out.print("input date (dd/MM/yyyy hh:mm): ");
                 String dateSTR = sc.nextLine();
 
                 try {
-                    date = du.ddMMyyyyHHm.parse(dateSTR);
+                    date = DateUtil.dateMinute.parse(dateSTR);
                 } catch (ParseException e) {
                     System.out.println("Error: wrong date format! example (12/08/2000 12:30)");
                     return;
                 }
 
                 if (storage.dateChecker(date)) {
-                    Patient patient = new Patient(patientId, patientName, patientSurname, patientEmail, patientPhoneNumber, patientsDoctor, date);
+                    Patient patient = new Patient(patientId, patientName, patientSurname, patientEmail,
+                            patientPhoneNumber, patientsDoctor, date);
                     storage.addPerson(patient);
                     System.out.println("Patient successfully registered");
                 } else {
                     System.out.println("This date is already booked");
                 }
+
             }
         }
     }
@@ -136,6 +161,7 @@ public class Demo implements Commands {
             System.out.println("Doctor with id " + changeId + " found!");
             Doctor doctorDataChange = storage.getDoctorById(changeId);
             System.out.println(doctorDataChange);
+
             System.out.println("Now input new data!");
             System.out.print("name: ");
             String newName = sc.nextLine();
@@ -144,12 +170,12 @@ public class Demo implements Commands {
             System.out.print("email: ");
             String newEmail = sc.nextLine();
 
-            boolean ec = emailChecker(newEmail);
+            boolean ec = isValidEmail(newEmail);
             while (!ec) {
                 System.out.println("wrong email format, try again! (ex@mail.ru, ex@gmail.com)");
                 System.out.print("email: ");
                 newEmail = sc.nextLine();
-                ec = emailChecker(newEmail);
+                ec = isValidEmail(newEmail);
             }
 
             System.out.print("phone number: ");
@@ -167,11 +193,19 @@ public class Demo implements Commands {
             doctorDataChange.setSurname(newSurname);
             doctorDataChange.setEmail(newEmail);
             doctorDataChange.setPhoneNumber(newPhoneNumber);
-            doctorDataChange.setProfession(Profession.valueOf(newProfession));
+
+            try {
+                doctorDataChange.setProfession(Profession.valueOf(newProfession));
+            } catch (IllegalArgumentException e) {
+                System.out.println("Profession " + newProfession + " not found!");
+                return;
+            }
 
             System.out.println("Doctor information successfully changed");
             System.out.println(doctorDataChange);
 
+        }else{
+            System.out.println("Doctor with id " + changeId + " not found!");
         }
     }
 
@@ -181,20 +215,22 @@ public class Demo implements Commands {
         for (Profession value : values) {
             System.out.println(value);
         }
+
         System.out.print("input profession: ");
         String searchProfession = sc.nextLine();
-        Doctor dc = null;
+        Doctor dc;
 
         try {
             dc = storage.searchDoctorByProfession(Profession.valueOf(searchProfession));
-        }catch(IllegalArgumentException iae){
-            System.out.println("Wrong profession!");
+        } catch (IllegalArgumentException iae) {
+            System.out.println("Profession " + searchProfession + " not found!");
+            return;
         }
 
-        if (dc != null) {
+        if(dc != null) {
             System.out.println(dc);
-        } else {
-            System.out.println("Profession " + searchProfession + " not found!");
+        }else{
+            System.out.println("There is no doctor with this profession!");
         }
     }
 
@@ -207,13 +243,15 @@ public class Demo implements Commands {
         String doctorSurname = sc.nextLine();
         System.out.print("input email: ");
         String doctorEmail = sc.nextLine();
-        boolean ec = emailChecker(doctorEmail);
+
+        boolean ec = isValidEmail(doctorEmail);
         while (!ec) {
             System.out.println("wrong email format, try again! (ex@mail.ru, ex@gmail.com");
             System.out.print("input email: ");
             doctorEmail = sc.nextLine();
-            ec = emailChecker(doctorEmail);
+            ec = isValidEmail(doctorEmail);
         }
+
         System.out.print("input phone number: ");
         String doctorPhoneNumber = sc.nextLine();
 
@@ -225,13 +263,19 @@ public class Demo implements Commands {
         System.out.print("input profession: ");
         String doctorProfession = sc.nextLine();
 
-        Doctor newDoctor = new Doctor(doctorId, doctorName, doctorSurname, doctorEmail, doctorPhoneNumber, Profession.valueOf(doctorProfession));
+        Doctor newDoctor;
+
+        try {
+            newDoctor = new Doctor(doctorId, doctorName, doctorSurname, doctorEmail,
+                    doctorPhoneNumber, Profession.valueOf(doctorProfession));
+        } catch (IllegalArgumentException e) {
+            System.out.println("No profession named " + doctorProfession + " found");
+            return;
+        }
+
         storage.addPerson(newDoctor);
         System.out.println("Doctor successfully added!");
-    }
 
-    private static boolean emailChecker(String email) {
-        return email.contains("@") && (email.endsWith(".ru") || email.endsWith(".com"));
     }
 
 }
